+++ desktopApp/src/jvmMain/kotlin/com/vitorpamplona/amethyst/desktop/ui/chats/DmBroadcastBanner.kt
@@ -18,7 +18,7 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.commons.ui.chat
+package com.vitorpamplona.amethyst.desktop.ui.chats
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.expandVertically
